requirements.txt
requests==2.25.1
beautifulsoup4==4.9.3
urllib3==1.26.5
pyjwt==2.4.0